Add menu file
