+++ main.rb
@@ -1,4 +1,6 @@
 require_relative 'app'
+require_relative 'menu'
+
 def main
   app = App.new
   puts 'Welcome to School Library App!'
@@ -14,19 +16,6 @@ def main
       process_option(option, app)
     end
   end
-end
-
-def show_menu
-  puts 'Please choose an option by entering a number:'
-  puts '1 - List all books'
-  puts '2 - List all people'
-  puts '3 - Create a person'
-  puts '4 - Create a book'
-  puts '5 - Create a rental'
-  puts '6 - List all rentals for a given person id'
-  puts '7 - Exit'
-  puts ' '
-  puts 'Enter number: '
 end
 
 def process_option(option, app)
